autocompleteの実装とテストを分割
