+++ src/lib/autocomplete/generate.spec.ts
@@ -1,9 +1,6 @@
-import {
-  generateAutocompleteOptions,
-  isOptionEqualToWord,
-} from './autocomplete';
-import { SearchType } from './search-target';
-import { Ship } from './ship';
+import { SearchType } from '../search-target';
+import { Ship } from '../ship';
+import { generateAutocompleteOptions } from './generate';
 
 describe('generateAutocompleteOptions', () => {
   const ships: Ship[] = [
@@ -100,36 +97,6 @@ describe('generateAutocompleteOptions', () => {
             false
           )
         ).toEqual(['x-ray']);
-      });
-    });
-  });
-});
-
-describe('isOptionEqualToWord', () => {
-  describe('最初のマイナス(-)を除いてラベルとワードが一致する場合', () => {
-    describe.each`
-      option     | word
-      ${'test'}  | ${'test'}
-      ${'test'}  | ${'-test'}
-      ${'-test'} | ${'test'}
-      ${'-test'} | ${'-test'}
-    `('タグラベルが $label でワードが $word の場合', ({ option, word }) => {
-      it('同じと判定される', () => {
-        expect(isOptionEqualToWord(option, word)).toBeTruthy();
-      });
-    });
-  });
-
-  describe('最初のマイナス(-)を除いてもラベルとワードが一致しない場合', () => {
-    describe.each`
-      option      | word
-      ${'label'}  | ${'word'}
-      ${'label'}  | ${'-word'}
-      ${'-label'} | ${'word'}
-      ${'-label'} | ${'-word'}
-    `('タグラベルが $label でワードが $word の場合', ({ option, word }) => {
-      it('違うと判定される', () => {
-        expect(isOptionEqualToWord(option, word)).toBeFalsy();
       });
     });
   });
