+++ src/lib/autocomplete/generate.ts
@@ -1,5 +1,5 @@
-import { SearchTarget, SearchType } from './search-target';
-import { Ship } from './ship';
+import { SearchTarget, SearchType } from '../search-target';
+import { Ship } from '../ship';
 
 const uniqueAndSortTagLabels = (tagLabels: string[]): string[] => {
   return Array.from(new Set(tagLabels)).sort();
@@ -22,19 +22,4 @@ export const generateAutocompleteOptions = (
     case SearchType.NAME:
       return shipsShown.map(({ name }) => name);
   }
-};
-
-const wordWithoutFirstMinus = (word: string): string => {
-  return word.match(/-?(.*)/)[1];
-};
-
-export const isOptionEqualToWord = (
-  // 補完対象から選択した場合AutocompleteOption、最後まで自分で入力した場合はstringになる
-  option: string,
-  word: string
-): boolean => {
-  // マイナス検索しているラベルやマイナス検索中のプラスラベルを除外する
-  const pureWord = wordWithoutFirstMinus(word);
-  const pureLabel = wordWithoutFirstMinus(option);
-  return pureWord === pureLabel;
 };
