+++ src/lib/autocomplete/index.ts
@@ -1,0 +1,2 @@
+export * from './compare';
+export * from './generate';
